Add skeleton ocaml hvbench
This includes the build runes, but no actual code.

Note a subtle difference with the c/ version: this one installs the
OCaml dependencies in the container but then builds from the current
directory as a shared volume. This is good for incremental development,
but a release build would need to be preceeded by `make clean`

Signed-off-by: David Scott <dave.scott@docker.com>